backend/requirements.txt
django>=6.0
djangorestframework>=3.15
django-cors-headers>=4.3
